# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NetMon — Network Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hosts/{id}/agent/latest:
    get:
      operationId: getHostAgentLatest
      tags: [hosts]
      summary: Latest windows_exporter metrics sample for a host
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest sample
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostAgentMetric"
        "404":
          description: No metrics collected yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{id}/agent/refresh:
    post:
      operationId: refreshHostAgent
      tags: [hosts]
      summary: Force an immediate windows_exporter scrape for a host and return the updated sample
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freshly collected sample
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostAgentMetric"
        "404":
          description: Host not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hosts/{id}/agent/test:
    get:
      operationId: testHostAgent
      tags: [hosts]
      summary: Test the windows_exporter connection for a host (single scrape)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: ip
          in: query
          required: false
          schema:
            type: string
        - name: port
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostAgentTestResult"
...
  /llm-models:
    get:
      operationId: listLlmModels
      tags: [llm-models]
      summary: List models configured for every provider
      responses:
        "200":
          description: All models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LlmModel"
    post:
      operationId: createLlmModel
      tags: [llm-models]
      summary: Add a model to a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmModelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmModel"
  /llm-models/{id}:
    put:
      operationId: updateLlmModel
      tags: [llm-models]
      summary: Update a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmModelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmModel"
    delete:
      operationId: deleteLlmModel
      tags: [llm-models]
      summary: Remove a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hosts/discover:
    post:
      operationId: discoverHosts
      tags: [hosts]
      summary: Scan network range for hosts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostDiscoveryInput"
      responses:
        "200":
          description: Discovered hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveredHost"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        canAccessAiTerminal:
          type: boolean
      required: [id, username, role, canAccessAiTerminal]
...
    HostAgentProcess:
      type: object
      properties:
        name:
          type: string
        running:
          type: boolean
        cpuPercent:
          type: ["number", "null"]
        memBytes:
          type: ["number", "null"]
      required: [name, running]
    HostAgentMetric:
      type: object
      properties:
        id:
          type: integer
        hostId:
          type: integer
        collectedAt:
          type: string
        cpuPercent:
          type: ["number", "null"]
        cpuModel:
          type: ["string", "null"]
        memUsedBytes:
          type: ["number", "null"]
        memTotalBytes:
          type: ["number", "null"]
        uptimeSeconds:
          type: ["number", "null"]
        osVersion:
          type: ["string", "null"]
        scrapeOk:
          type: boolean
        processes:
          type: array
          items:
            $ref: "#/components/schemas/HostAgentProcess"
      required: [id, hostId, collectedAt, scrapeOk, processes]
    HostAgentTestResult:
      type: object
      properties:
        ok:
          type: boolean
        url:
          type: string
        statusCode:
          type: ["integer", "null"]
        error:
          type: ["string", "null"]
        sample:
          type: ["object", "null"]
      required: [ok, url]
...
    LlmModel:
      type: object
      properties:
        id:
          type: integer
        provider:
          type: string
          enum: [claude, gpt, gemini, deepseek, groq]
        modelId:
          type: string
        label:
          type: string
        enabled:
          type: boolean
        orderIndex:
          type: integer
      required: [id, provider, modelId, label, enabled, orderIndex]
    LlmModelInput:
      type: object
      properties:
        provider:
          type: string
          enum: [claude, gpt, gemini, deepseek, groq]
        modelId:
          type: string
        label:
          type: string
        enabled:
          type: boolean
        orderIndex:
          type: integer
      required: [provider, modelId, label]
    LlmModelUpdate:
      type: object
      properties:
        modelId:
          type: ["string", "null"]
        label:
          type: ["string", "null"]
        enabled:
          type: ["boolean", "null"]
        orderIndex:
          type: ["integer", "null"]
    HostDiscoveryInput:
      type: object
      properties:
        startIp:
          type: string
        endIp:
          type: string
      required: [startIp, endIp]
    DiscoveredHost:
      type: object
      properties:
        ipAddress:
          type: string
        reachable:
          type: boolean
      required: [ipAddress, reachable]
...